refactor(types): split notifications.ts into 3 per-entity files
